Add support for scoped package (Fix #272) (#275)
* Add supports for scoped package (Fix #272)

* Streamline option `githubAccount`

* change option name from `github-account` to `githubAccount` in
git/index.js (githubAccount is used everywhere else, and the result is
the same)
* change option description in readme/index.js to match the description
in app/index.js

* Add option `repositoryName`

Add the option `repositoryName`, being the ‘Name of the GitHub
repository’.

* fixup! Add supports for scoped package (Fix #272)

* Update package-lock.json

* fixup! fixup! Add supports for scoped package (Fix #272)

Fix on node@4.x.x

* Bump dependencies (inquirer-npm-name@3)

* Use package scope as default `githubUser`

If a scoped name is supplied, use the scope part (after the `@` and
before the first `/`) as default value for `githubUser`, even ignoring
the `authorEmail` which is otherwise used to find the default value.

* Drop `name`-option from git-generator

* Only write `repository` into pkg if available

Only write `repository` to the package file if both `githubAccount`
and `repositoryName` are passed to the `git`-generator. Re-read the
package before adding the `origin`-remote to repository in case some
other generator has modified it.

* fixup! Use package scope as default `githubUser`

* fixup! Only write `repository` into pkg if available

* fixup! fixup! Only write `repository` into pkg if available

* Remove _readPkg side effects from git subgenerator

* Remove the side effects of _getModuleNameParts

Have `_getModuleNameParts` return an object with the parts of the module
name instead of directly writing it to `props`.